# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Simple Inventory API",
    "description" : "This is a simple API",
    "contact" : {
      "email" : "you@your-company.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.2"
  },
  "servers" : [ {
    "url" : "https://api.example2.com",
    "description" : "New Server 2"
  }, {
    "url" : "https://api.example3.com",
    "description" : "New Server 3"
  } ],
  "tags" : [ {
    "name" : "admins",
    "description" : "Secured Admin-only calls"
  }, {
    "name" : "developers",
    "description" : "Operations available to regular developers"
  }, {
    "name" : "New Tag 3"
  } ],
  "paths" : {
    "/inventory" : {
      "get" : {
        "tags" : [ "developers" ],
        "summary" : "searches inventory",
        "description" : "By passing in the appropriate options, you can search for\navailable inventory in the system\n",
        "operationId" : "searchInventory",
        "parameters" : [ {
          "name" : "searchString",
          "in" : "query",
          "description" : "pass an optional search string for looking up inventory",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "maximum" : 50.0,
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/InventoryItem"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      },
      "post" : {
        "tags" : [ "admins" ],
        "summary" : "adds an inventory item",
        "description" : "Adds an item to the system",
        "operationId" : "addInventory",
        "requestBody" : {
          "description" : "Inventory item to add",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/InventoryItem"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "item created"
          },
          "400" : {
            "description" : "invalid input, object invalid"
          },
          "409" : {
            "description" : "an existing item already exists"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "InventoryItem" : {
        "required" : [ "id", "manufacturer", "name", "releaseDate" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
          },
          "name" : {
            "type" : "string",
            "example" : "Widget Adapter"
          },
          "releaseDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "manufacturer" : {
            "$ref" : "#/components/schemas/Manufacturer"
          }
        },
        "example" : {
          "releaseDate" : "2000-01-23T04:56:07.000+00:00",
          "name" : "Widget Adapter",
          "id" : "d290f1ee-6c54-4b01-90e6-d701748f0851",
          "manufacturer" : {
            "phone" : "408-867-5309",
            "name" : "ACME Corporation",
            "homePage" : "https://www.acme-corp.com"
          }
        }
      },
      "Manufacturer" : {
        "required" : [ "name" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "example" : "ACME Corporation"
          },
          "homePage" : {
            "type" : "string",
            "format" : "url",
            "example" : "https://www.acme-corp.com"
          },
          "phone" : {
            "type" : "string",
            "example" : "408-867-5309"
          }
        },
        "example" : {
          "phone" : "408-867-5309",
          "name" : "ACME Corporation",
          "homePage" : "https://www.acme-corp.com"
        }
      }
    }
  }
}